WIP: adding date fn with locale code split
