+++ src/components/DisplayDate/index.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+interface DisplayDateProps {
+  date: Date;
+  format?: string;
+}
+
+export default function DisplayDate({ date, format = 'PP' }: DisplayDateProps) {
+  
+
+  return ()
+}
