+++ src/lib/date-i18n.ts
@@ -1,0 +1,8 @@
+import { ValidLocale } from './i18n'
+
+const getDateFnLocale = async (locale: ValidLocale) => {
+  const datefnLocale = await import(`date-fns/locale/${locale}`)
+  console.log(datefnLocale);
+  
+  return datefnLocale;
+}
