+++ src/lib/i18n.ts
@@ -12,11 +12,6 @@ export type TranslatorFn = (
 
 export type Translator = TranslatorFn & { locale?: string };
 
-//const dictionaries: Record<ValidLocale, any> = {
-    //en: import('./dictionaries/en.json').then((module) => module.default),
-    //es: import('./dictionaries/es.json').then((module) => module.default),
-//}
-
 export const getLocaleFrom = (pathname: string): string => {
     const [, locale] = pathname.toLowerCase().split('/');
     return locale;
@@ -44,9 +39,7 @@ export const getTranslatorFrom = (dictionary: any): Translator => {
 }
 
 export const getTranslator = async (locale: ValidLocale) => {
-    const dictionary = await import(`./dictionaries/${locale}.json`);
-    //const dictionary = await dictionaries[locale]();
+    const dictionary = await import(`@src/dictionaries/${locale}.json`);
 
     return getTranslatorFrom(dictionary);
 }
-
